[TM-3552] split into components
