+++ src/pages/reports/site-report/tabs/Details.tsx
@@ -1,9 +1,6 @@
-import { useT } from "@transifex/react";
 import { Dictionary } from "lodash";
 import { FC } from "react";
 
-import EmptyState from "@/components/elements/EmptyState/EmptyState";
-import { IconNames } from "@/components/extensive/Icon/Icon";
 import SharedDetails from "@/components/extensive/PageElements/PageContent/components/sharedDetails";
 import PageContent from "@/components/extensive/PageElements/PageContent/PageContent";
 import { FormStepWithValidation } from "@/components/extensive/WizardForm/useFormStepsWithValidation";
@@ -12,6 +9,7 @@ import WizardFormProvider from "@/context/wizardForm.provider";
 import { SiteReportFullDto } from "@/generated/v3/entityService/entityServiceSchemas";
 import { useEntityFormSetup } from "@/hooks/useEntityFormSetup";
 import { useProjectOrgFormData } from "@/hooks/useProjectOrgFormData";
+import NothingToReportEmptyState from "@/pages/reports/site-report/components/NothingToReportEmptyState";
 
 interface SiteReportDetailsTabProps {
   report: SiteReportFullDto;
@@ -39,7 +37,6 @@ const SharedDetailsStep: FC<SharedDetailsStepProps> = ({ step, formValues, repor
 );
 
 const SiteReportDetailsTab: FC<SiteReportDetailsTabProps> = ({ report }) => {
-  const t = useT();
   const { steps, defaultValues, fieldsProvider, isFormLoading, providerLoaded } = useEntityFormSetup(
     "site-reports",
     report.uuid
@@ -51,13 +48,7 @@ const SiteReportDetailsTab: FC<SiteReportDetailsTabProps> = ({ report }) => {
   if (report.nothingToReport) {
     return (
       <PageContent className="gap-2 bg-theme-neutral-100 sm:px-32">
-        <EmptyState
-          iconProps={{ name: IconNames.DOCUMENT_CIRCLE, className: "fill-success" }}
-          title={t("Nothing to report")}
-          subtitle={t(
-            "You've marked this report as 'Nothing to Report,' indicating there are no updates for this site report. If you wish to add information to this report, please use the edit button."
-          )}
-        />
+        <NothingToReportEmptyState />
       </PageContent>
     );
   }
